backend/requirements.txt
fastapi
pyttsx3
pydub
moviepy
groq
ffmpeg
python_dotenv